ztp: Add siteConfigHelper_test
